# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: FoodBucket
  description: This is a server API for FoodBucket project
  termsOfService: http://foodbucket.com/terms/
  contact:
    name: Foodbucket API support team
    url: http://foodbucket.com/support
    email: support@foodbucket.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
# during dev, should point to your local machine
host: localhost:3000
...
basePath: /api
schemes:
  # tip: remove http to make production-grade
  - http
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
...
  /promotion:
    get:
      x-swagger-router-controller: Promotion
      operationId: getPromotion
      tags:
      - "promotion"
      produces:
      - application/json
      parameters:
        - name: offset
          in: query
          description: start position for quering from DB
          required: true
          type: integer
          format: int32
          minimum: 0
          default: 0
        - name: limit
          in: query
          description: number of items to query from DB
          required: true
          type: integer
          format: int32
          minimum: 0
          maximum: 28
          default: 9
        - name: isPromotion
          in: query
          description: returns promotion products
          required: false
          default: true
          type: boolean
      responses:
        200:
          description: Successful products response
          schema:
            type: array
            items:
              $ref: "#/definitions/Product"
        default:
          description: invalid input
          schema:
            $ref: "#/definitions/Error"
...
  /auth/validation:
      get:
        x-swagger-router-controller: Auth
        operationId: validation
        tags:
        - "auth"
        consumes:
        - application/json
        summary: Validation operation
        description: |
          This endpoint allows to register.
        parameters:
        - in: header
          name: X-MY-JWT
          description: Validation object
          type: string
          required: true
        responses:
          200:
            description: Existing registration was found and updated
            schema:
              $ref: '#/definitions/Validation'
          201:
            description: Accepted
            schema:
              $ref: '#/definitions/Validation'
          400:
            description: Invalid input
...
definitions:
...
  Product:
    type: object
    required:
      - title
      - description
      - price
    properties:
      productId:
        type: integer
        format: int64
      title:
        type: string
      description:
        type: string
      image:
        type: string
      price:
        type: integer
        format: int64
        minimum: 1
        default: 1
      category:
        type: string
      status:
        type: boolean
      recommended:
        type: [string]
      discount:
        type: integer
        format: int64
      promotions:
        type: boolean
      caloricity:
        type: integer
        format: int64
      servingSize:
        type: integer
      difficulty:
        type: string
      spiceLevel:
        type: string
      ingredients:
        type: array
        items:
          $ref: "#/definitions/IngredientsList"
  IngredientsList:
    type: object
    properties:
      ingredientId:
        type: integer
        format: int64
      ingredientName:
        type: string
      quantity:
        type: integer
        format: int64
      measure:
        type: string
...
  Validation:
    type: object
    properties:
      isValid:
        type: boolean
  Error:
    type: object
    required:
      - code
      - message
    properties:
      code:
        type: integer
        format: int32
      message:
        type: string